Initial file.
Reads all flights from www.onlinecontest.org that started in the Alps. For now, it can read the overview pages of the years that are put online. It puts the data into a table, so it can be analyzed later. This table also contains the links to the detail pages. Other than the overview pages, they are javascript rendered and I need to use another approach. This will happen in one of the following versions.